Feat: Discovery con pines configurados, documentacion y updateHIDData implementada

+++ Libs/Inc/Encoder_types.h
@@ -15,8 +15,8 @@
 
 typedef struct {
     int32_t  rotations;      // 4 bytes - Cantidad de vueltas completas
-    float    total_degrees;  // 4 bytes - Posición absoluta en grados
-    float    velocity_dps;   // 4 bytes - Velocidad en grados/segundo
+    int32_t  total_degrees;  // 4 bytes - Posición absoluta en grados
+    int32_t  velocity_dps;   // 4 bytes - Velocidad en grados/segundo
     uint16_t raw_angle;      // 2 bytes - Valor crudo del registro (0-4095)
     uint8_t  buffer[2];      // 2 bytes - Buffer para el DMA (I2C)
     uint8_t  status;         // 1 byte  - Flags de error o estado del sensor
